[fix] #61 develop merge conflict fix
