+++ src/main/java/org/umc/travlocksserver/domain/vlock/service/command/VlockCommandService.java
@@ -15,6 +15,7 @@ import org.umc.travlocksserver.domain.member.repository.MemberRepository;
 import org.umc.travlocksserver.domain.vlock.code.VlockCategoryErrorCode;
 import org.umc.travlocksserver.domain.vlock.code.VlockErrorCode;
 import org.umc.travlocksserver.domain.vlock.dto.request.VlockRequestDTO;
+import org.umc.travlocksserver.domain.vlock.dto.request.VlockUpdateRequestDTO;
 import org.umc.travlocksserver.domain.vlock.dto.response.VlockResponseDTO;
 import org.umc.travlocksserver.domain.vlock.entity.Vlock;
 import org.umc.travlocksserver.domain.vlock.entity.VlockCategory;
@@ -40,14 +41,9 @@ public class VlockCommandService {
 	private final CityQueryService cityQueryService;
 
 	public VlockResponseDTO createVlock(Long memberId, VlockRequestDTO request) {
-		Member member = memberRepository.findById(memberId)
-			.orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
-
-		VlockCategory category = vlockCategoryRepository.findById(request.categoryId())
-			.orElseThrow(() -> new VlockException(VlockErrorCode.CATEGORY_NOT_FOUND));
-
-		City city = cityRepository.findWithRegionById(request.cityId())
-			.orElseThrow(() -> new CityException(CityErrorCode.CITY_NOT_FOUND));
+		Member member = getMember(memberId);
+		VlockCategory category = getCategory(request.categoryId());
+		City city = getCity(request.cityId());
 
 		Vlock newVlock = Vlock.create(
 			category,
@@ -63,6 +59,30 @@ public class VlockCommandService {
 		Vlock savedVlock = vlockRepository.save(newVlock);
 
 		return VlockResponseDTO.from(savedVlock);
+	}
+
+	public VlockResponseDTO updateVlock(Long memberId, Long vlockId, VlockUpdateRequestDTO request) {
+		validateMemberExists(memberId);
+
+		Vlock vlock = getOwnedVlock(memberId, vlockId);
+
+		VlockCategory category = getCategory(request.categoryId());
+		City city = getCity(request.cityId());
+
+		vlock.update(
+			category,
+			city,
+			request.name(),
+			request.latitude(),
+			request.longitude(),
+			request.address(),
+			request.memo(),
+			request.coverImgUrl(),
+			request.linkUrl(),
+			request.isPublic()
+		);
+
+		return VlockResponseDTO.from(vlock);
 	}
 
     // ⚪ 외부(카카오맵) API를 통해 블록을 삽입하는 메서드 (추천시 블록에 데이터가 너무 적을 경우 사용)
@@ -104,5 +124,36 @@ public class VlockCommandService {
         return vlockCategoryQueryService.getByName(name)
                 .orElseGet(() -> vlockCategoryQueryService.getByName("기타")
                         .orElseThrow(() -> new VlockCategoryException(VlockCategoryErrorCode.DEFAULT_VLOCK_CATEGORY_NOT_FOUND)));
-    };
+    }
+
+	private void validateMemberExists(Long memberId) {
+		if (!memberRepository.existsById(memberId)) {
+			throw new MemberException(MemberErrorCode.MEMBER_NOT_FOUND);
+		}
+	}
+
+	private Member getMember(Long memberId) {
+		return memberRepository.findById(memberId)
+			.orElseThrow(() -> new MemberException(MemberErrorCode.MEMBER_NOT_FOUND));
+	}
+
+	private VlockCategory getCategory(Long categoryId) {
+		return vlockCategoryRepository.findById(categoryId)
+			.orElseThrow(() -> new VlockException(VlockErrorCode.CATEGORY_NOT_FOUND));
+	}
+
+	private City getCity(Long cityId) {
+		return cityRepository.findWithRegionById(cityId)
+			.orElseThrow(() -> new CityException(CityErrorCode.CITY_NOT_FOUND));
+	}
+
+	private Vlock getOwnedVlock(Long memberId, Long vlockId) {
+		Vlock vlock = vlockRepository.findByIdAndDeletedAtIsNull(vlockId)
+			.orElseThrow(() -> new VlockException(VlockErrorCode.VLOCK_NOT_FOUND));
+
+		if (!vlock.isOwnedBy(memberId)) {
+			throw new VlockException(VlockErrorCode.VLOCK_FORBIDDEN);
+		}
+		return vlock;
+	}
 }
